Introduced the interface Machine that Host and Vm interfaces extend from.
- This interface provides common methods that existed in both Host and Vm interfaces,
  reducing code and documentation duplication.

+++ cloudsim-plus/src/main/java/org/cloudbus/cloudsim/datacenters/DatacenterCharacteristics.java
@@ -147,7 +147,7 @@ public interface DatacenterCharacteristics extends Identificable {
      * @pre $none
      * @post $result >= 0
      */
-    long getMips();
+    double getMips();
 
     /**
      * Gets Millions Instructions Per Second (MIPS) Rating of a Processing
